requirements.txt
fastapi
uvicorn
openai
pydantic
pillow
beautifulsoup4
requests
